Add test for no files selected
This was a way bigger pain than it needed to be. Alerts may or may not
show up in HtmlUnit in their own threads, which makes them very
difficult to respond to in the main thread. Hence the current
implementation just catches the exception for the missing KBase login
form and checks the alert list rather than trying to stop execution at
the point of the alert.