Add test for JGI-side shock node memoization
JGI keeps a file->shock node map and doesn't transfer the file if 1)
they have a node for the file and 2) the node still exists in shock.

This test pushes a file, thus creating a node, deletes the node in
shock, and then repushes to check that the file is transferred to shock
again.