Update tests to new working targets.
Filed JGI bugs for the tests that failed due to the SDM server returning
400s to the JSF front end. All tests now have the possibility of
passing, although they've been very unstable with lots of timeouts and
html structure errors. Not sure yet if this is just server instability
or whether the html structure changed and the tests need to be adjusted.